interfaccia grafica che funziona e calcola il costo corrrettamente, manca quella diamministratore e addetto e manca di considerare gli abbonamenti nel calcolo del costo

+++ src/main/java/com/parkHour/view/interfacciaVeicolo/ViewUscita.java
@@ -36,7 +36,7 @@ public class ViewUscita {
         else {
             abbonamento.setText("No");
         }
-        durataSosta.setText(minuti/60 + " ore "+ minuti%60+" minuti ");
+        durataSosta.setText((int)minuti/60 + " ore "+ (int)minuti%60+" minuti ");
         importoDaPagare=daPagareV;
         daPagare.setText(String. format("%.2f", daPagareV)+" €");
         versato.setText("0,00 €");
